api/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
pytest
requests
scikit-learn
